fix: refactoring in naming and code in Form component
