+++ src/pages/Login/index.tsx
@@ -1,8 +1,8 @@
 import { FieldValues } from 'react-hook-form';
 
 import { Button } from '@/components/Button/Button.tsx';
-import { FormWrapper } from '@/components/Form/FormWrapper';
-import { Form } from '@/components/Form/FormWrapper/style.ts';
+import { FormWrapper } from '@/components/Form';
+import { Form } from '@/components/Form/style.ts';
 import { Input } from '@/components/Input/Input.tsx';
 import { PATH } from '@/constants/path.ts';
 import { useFormHandler } from '@/hooks/useFormHandler.ts';
